+++ operator/api/v1beta1/lokistack_types.go
@@ -369,8 +369,8 @@ type ObjectStorageSecretSpec struct {
 	Name string `json:"name"`
 }
 
-// ObjectStorageSchemaMap defines the type for mapping a schema version with a date
-type ObjectStorageSchemaMap map[StorageSchemaEffectiveDate]ObjectStorageSchemaVersion
+// objectStorageSchemaMap defines the type for mapping a schema version with a date
+type objectStorageSchemaMap map[StorageSchemaEffectiveDate]ObjectStorageSchemaVersion
 
 // ObjectStorageSchemaVersion defines the storage schema version which will be
 // used with the Loki cluster.