+++ operator/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Main
 
-- [6411](https://github.com/grafana/loki/pull/6411) **Red-GV**: Add validation cases to lokistacks webhook
+- [6411](https://github.com/grafana/loki/pull/6411) **Red-GV**: Extend schema validation in LokiStack webhook
 - [6224](https://github.com/grafana/loki/pull/6224) **periklis**: Add support for GRPC over TLS for Loki components
 - [5952](https://github.com/grafana/loki/pull/5952) **Red-GV**: Add api to change storage schema version
 - [6363](https://github.com/grafana/loki/pull/6363) **periklis**: Allow optional installation of webhooks (Kind)